+++ src/main/java/thedarkcolour/exdeorum/compat/ModIds.java
@@ -41,4 +41,5 @@ public class ModIds {
     public static final String AETHER = "aether";
     public static final String BLUE_SKIES = "blue_skies";
     public static final String PAMS_HARVESTCRAFT_CROPS = "pamhc2crops";
+    public static final String NUCLEARCRAFT_NEOTERIC = "nuclearcraft";
 }